Commit for the 1.1.0 .exe build.

+++ moveCompiled.py
@@ -1,6 +1,7 @@
 import shutil
 import os
 import Paths
+import DataProcessorGUI
 #from DataProcessor import Basic
 
 
@@ -9,21 +10,22 @@ import Paths
 for file in ["Updater.exe"]:
     shutil.copy(Paths.join(Paths.compiledUpdater, file), Paths.compiledMain)
 
-if os.path.exists(Paths.sourcePath):
-    shutil.rmtree(Paths.sourcePath)
+if not DataProcessorGUI.debug:
+    if os.path.exists(Paths.sourcePath):
+        shutil.rmtree(Paths.sourcePath)
 
-#os.mkdir(Paths.sourcePath)
-shutil.move(Paths.compiledMain, Paths.downloadPath)
+    #os.mkdir(Paths.sourcePath)
+    shutil.move(Paths.compiledMain, Paths.downloadPath)
 
 
-for folder in ["icons", "exe", "DataProcessor/Text"]:
-    shutil.copytree(folder, Paths.sourcePath+'/'+folder, dirs_exist_ok=True)
+    for folder in ["icons", "exe", "DataProcessor/Text"]:
+        shutil.copytree(folder, Paths.sourcePath+'/'+folder, dirs_exist_ok=True)
 
-shutil.copy(Paths.configFile, Paths.getConfigFile(Paths.sourcePath))
-for file in []:
-    shutil.copy(file, Paths.sourcePath + "/" + file)
+    shutil.copy(Paths.configFile, Paths.getConfigFile(Paths.sourcePath))
+    for file in []:
+        shutil.copy(file, Paths.sourcePath + "/" + file)
 
-for folder in ["dist", "build"]:
-    shutil.rmtree(folder)
+    for folder in ["dist", "build"]:
+        shutil.rmtree(folder)
